requirements.txt
python-dotenv
requests
pandas
pyarrow
fastapi
uvicorn
pytest
httpx
